Refine URC, TVD proofs, add compact support proof
For convinience, the old and new time steps are denoted
as n,n+1 respectively.
The URC proof was made simpler, considering all three cases
of initial data.
A first attempt to prove that the solution has compact support
was added.
The TVD proof was also refined to be more followable.